requirements_additions.txt
# Сессия #4 зависимости
aiogram>=3.0.0
aiohttp>=3.8.0

# Сессия #5 зависимости
psutil>=5.9.0
rich>=13.0.0
pyttsx3>=2.90
openpyxl>=3.1.0
requests-cache>=1.0.0
